requirements.txt
astroid==3.3.8
black==24.10.0
click==8.1.8
dill==0.3.9
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
picamera==1.13
pillow==11.1.0
platformdirs==4.3.6
pylint==3.3.3
tomlkit==0.13.2
